fix 64 bit compile issues.

+++ Classes/MGSplitCornersView.h
@@ -17,13 +17,13 @@ typedef enum _MGCornersPosition {
 
 @class MGSplitViewController;
 @interface MGSplitCornersView : UIView {
-	float cornerRadius;
+	CGFloat cornerRadius;
 	MGSplitViewController *__unsafe_unretained splitViewController;
 	MGCornersPosition cornersPosition;
 	UIColor *cornerBackgroundColor;
 }
 
-@property (nonatomic, assign) float cornerRadius;
+@property (nonatomic, assign) CGFloat cornerRadius;
 @property (nonatomic, unsafe_unretained) MGSplitViewController *splitViewController; // weak ref.
 @property (nonatomic, assign) MGCornersPosition cornersPosition; // don't change this manually; let the splitViewController manage it.
 @property (nonatomic, strong) UIColor *cornerBackgroundColor;
